[api] improve api endpoints (#198)
* [api] improve api endpoints

* [api] improve api endpoints

* [api] remove alert and alertuser

* merge with origin

* [fix] remove alert and alert user

* [fix] remove console.log

+++ Hack/Backend/controllers/UserController.js
@@ -6,7 +6,7 @@ const axios = require('axios');
 const config = require('config');
 const User = UserModels.User;
 
-// Get All Users.
+// Get All Users. - [DEPRECATED: The information is too sensitive to share with API consumers]
 exports.get_all_users = async (req, res) => {
   if (req.query.demo && req.query.demo == "true"){
     var User = UserModels.DemoUser;
@@ -30,6 +30,10 @@ exports.get_user_by_id = async (req, res) => {
     var User = UserModels.StressUser;
   }else{
     var User = UserModels.User;
+  }
+  if(req.params.id !== req.body.loggedInUser){
+
+    return res.status(403).send("User not authorized to access this endpoint with id: " + req.body.id);
   }
   const user = await User.findById(req.params.id);
   try {
@@ -121,6 +125,8 @@ exports.update_user = async (req, res) => {
     let change = await User.findOne({ _id: req.body._id })
     if (exists && exists._id != req.body._id) {
       return res.status(400).send("Username already exists ");
+    } else if(req.body._id !== req.body.loggedInUser){
+      return res.status(403).send("User not authorized to access this endpoint with id: " + req.body._id);
     }
     if (req.body.password) {
       if (req.body.password.length < 5) {
@@ -140,12 +146,17 @@ exports.update_user = async (req, res) => {
 exports.delete_user = async (req, res) => {
 
   try {
+    if(req.body._id !== req.body.loggedInUser){
+      return res.status(403).send("User not authorized to access this endpoint with id: " + req.body._id);
+    }
     const user = await User.findByIdAndDelete(req.body._id);
     if (!user) {
       return res.status(404).send("No item found");
     }
     res.status(201).send(user);
   } catch (err) {
+    console.log(err.toString());
+    
     res.status(500).send(err.toString());
   }
 };
